add yes no modal

+++ BlazorPWA2/Pages/Contacts/ContactsPage.razor.cs
@@ -1,5 +1,8 @@
 ﻿using BlazorBootstrap;
+using Blazored.Modal;
+using Blazored.Modal.Services;
 using BlazorPWA2.Interfaces;
+using BlazorPWA2.Modal;
 using BlazorPWA2.Model;
 using BlazorPWA2.Services;
 using Microsoft.AspNetCore.Components;
@@ -9,16 +12,17 @@ namespace BlazorPWA2.Pages.Contacts;
 public partial class ContactsPage : ComponentBase
 {
     [Inject] public IContactRepository ContactRepository { get; set; }
-    [Inject] ModalService modalService { get; set; }
+    [Inject] IModalService modalService { get; set; }
     [Inject] INavigationService navigationService { get; set; }
-    
+    [Inject] ILogger<ContactsPage> logger {get;set;}
+
 
     public List<Contact> ContactList { get; set; } = new();
 
     protected override async Task OnInitializedAsync()
     {
         await base.OnInitializedAsync();
-        if(ContactRepository != null)
+        if (ContactRepository != null)
         {
             ContactList = await ContactRepository.GetAsync();
         }
@@ -29,21 +33,26 @@ public partial class ContactsPage : ComponentBase
         navigationService.NavigateTo("/contacts/edit");
     }
 
-    private async void OnDeleteClicked(Contact? contact){
-        if(contact != null){
-            try{
-                await ContactRepository.DeleteAsync(contact.Id);
-                ContactList.Remove(contact);
-            }
-            catch (Exception ex){
-                var modalOption = new ModalOption{
-                    Title = "Kontakt konnte nicht gelöscht werden",
-                    Message = ex.Message,
-                    Type = ModalType.Danger
-                };
+    private async void OnDeleteClicked(Contact? contact)
+    {
+        if (contact != null)
+        {
+            var modal = modalService.Show(typeof(YesNoModal));
+            var result = await modal.Result;
 
-                await modalService.ShowAsync(modalOption);
+            if (result.Confirmed)
+            {
+                try
+                {
+                    await ContactRepository.DeleteAsync(contact.Id);
+                    ContactList.Remove(contact);
+                }
+                catch (Exception ex)
+                {
+                    logger.LogError(ex.Message);
+                }
             }
+
         }
     }
 }
